Refactored code to meet all Style Checks
